backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
whitenoise
celery
redis
django-redis
flower
